hdoc almost good
