+++ sources/tools/tools2.c
@@ -3,15 +3,28 @@
 /*                                                        :::      ::::::::   */
 /*   tools2.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tmejri <tmejri@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jthuysba <jthuysba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/25 14:24:17 by tmejri            #+#    #+#             */
-/*   Updated: 2023/05/26 16:16:01 by tmejri           ###   ########.fr       */
+/*   Updated: 2023/05/30 21:00:01 by jthuysba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "heredoc.h"
+
+// void	*ft_calloc(size_t n, size_t size)
+// {
+// 	int	*arr;
+
+// 	if (size && n > (size_t) - 1 / size)
+// 		return (0);
+// 	arr = (void *)malloc(n * size);
+// 	if (!arr)
+// 		return (NULL);
+// 	ft_bzero(arr, n * size);
+// 	return (arr);
+// }
 
 int	ft_strcmp(char *s1, char *s2)
 {
