+++ sources/we_exec/prep_cmd.c
@@ -6,7 +6,7 @@
 /*   By: jthuysba <jthuysba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/28 16:21:12 by jthuysba          #+#    #+#             */
-/*   Updated: 2023/05/30 19:14:27 by jthuysba         ###   ########.fr       */
+/*   Updated: 2023/05/30 19:52:47 by jthuysba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,42 +52,6 @@ t_list	**clone_to_op(t_list *token)
 		elem = elem->next;
 	}
 	return (lst);
-}
-
-/* Enregistre les redirections avec leur type et leur file associes
--> cat < file */
-int	handle_redir(t_cmd *cmd, t_list *elem)
-{
-	int	i;
-
-	cmd->nb_redir = nb_redir(*(cmd->token));
-	if (cmd->nb_redir == 0)
-		return (0);
-	cmd->redir = malloc(sizeof(int) * cmd->nb_redir);
-	if (!cmd->redir)
-		return (1);
-	cmd->file = malloc(sizeof(char *) * (cmd->nb_redir + 1));
-	if (!cmd->file)
-		return (1);
-	ft_memset(cmd->file, 0, sizeof(char) * cmd->nb_redir);
-	i = 0;
-	while (elem)
-	{
-		if (elem->type != WORD)
-		{
-			cmd->redir[i] = elem->type;
-			if (elem->next)
-				cmd->file[i] = ft_strdup(elem->next->content);
-			else
-				return (1);//WIP
-			if (!cmd->file[i])
-				return (1);
-			i++;
-		}
-		elem = elem->next;
-	}
-	// cmd->file[i] = ft_strdup(NULL);
-	return (0);
 }
 
 /* Verifie si la commande est valide
